[fix] KoinPicker 변경 및 gradient 추가
